Clean up codebase (#884)
### Description
Performed general code cleanup to improve readability and
maintainability.

### Key Changes
- Removed redundant code.
- Enhanced code readability by applying consistent formatting.

### Types of Changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation fix (typos, incorrect content, missing content,
etc.)
- [x] Code refactoring
- [ ] Merge Down

### Test Case
N/A (No functional changes introduced).

+++ tests/NeoServer.Domain.Tests/Creature/Creature/PlayerTest.cs
@@ -328,7 +328,7 @@ public class PlayerTest
 
         player.Death(null);
 
-        Assert.Equal(initialExp * 0.9, (double)player.Experience); // 10% loss
+        Assert.Equal(initialExp * 0.9, player.Experience); // 10% loss
         Assert.Equal(23, player.Level);
         Assert.False(player.IsPromoted);
     }
@@ -364,7 +364,7 @@ public class PlayerTest
 
         player.Death(null);
 
-        Assert.Equal(expectedExpAfterDeath, (double)player.Experience);
+        Assert.Equal(expectedExpAfterDeath, player.Experience);
         Assert.Equal(50, player.Level); // Should stay at level 50
         Assert.True(player.IsPromoted);
     }
